bump

+++ src/hmpps/clients/github.py
@@ -4,6 +4,7 @@ import json
 import yaml
 import jwt
 import sys
+import os
 from time import sleep
 from github import Auth, Github
 from github import GithubException
@@ -19,40 +20,61 @@ from hmpps.services.job_log_handling import (
 
 
 class GithubSession:
-  def __init__(self, params):
+  def __init__(
+    self,
+    org: str = 'ministryofjustice',
+    app_id: str = '',
+    app_installation_id: str = '',
+    app_private_key: str = '',
+    access_token: str = '',
+    bootstrap_repo: str = '',
+  ):
+    self.org_name = org
+
+    # Define defaults
+    self.app_id = app_id or os.getenv('GITHUB_APP_ID', '')
+    self.app_installation_id = app_installation_id or os.getenv(
+      'GITHUB_APP_INSTALLATION_ID', ''
+    )
+    self.app_private_key = app_private_key or os.getenv('GITHUB_APP_PRIVATE_KEY', '')
+    self.access_token = access_token or os.getenv('GITHUB_ACCESS_TOKEN', '')
+    self.bootstrap_repo_name = (
+      bootstrap_repo
+      or os.environ.get('GITHUB_BOOTSTRAP_REPO', '')
+      or 'hmpps-project-bootstrap'
+    )
+
     # Don't progress if there's no private key or access token
-    if not params.get('app_private_key') and not params.get('github_access_token'):
+    if not self.app_private_key and not self.access_token:
       log_error(
         'app_private_key or github_access_token are needed to initiate a Github Session'
       )
       sys.exit(1)
 
     # set the parameters depending on authentication type
-    if params.get('app_private_key'):
-      self.private_key = b64decode(params.get('app_private_key')).decode('ascii')
-      self.app_id = params.get('app_id')
-      self.app_installation_id = params['app_installation_id']
+    if self.app_private_key:
+      log_debug('Using private key provided to authenticate')
+      self.private_key = b64decode(self.app_private_key).decode('ascii')
     else:
+      log_debug('Using access token provided to authenticate')
       self.private_key = self.app_id = self.app_installation_id = ''
-      self.rest_token = params.get('github_access_token')
-
-    self.org_name = params.get('org', 'ministryofjustice')
+      self.rest_token = self.access_token
 
     # Create a session with a private key
     self.auth()
-    try:
-      rate_limit = self.session.get_rate_limit()
-      self.core_rate_limit = rate_limit.resources.core
-      log_info(f'Github API - rate limit: {rate_limit}')
-    except Exception as e:
-      log_error(f'Unable to get github rate limit - {e}')
+
+    if self.session:
+      try:
+        rate_limit = self.session.get_rate_limit()
+        self.core_rate_limit = rate_limit.resources.core
+        log_info(f'Github API - rate limit: {rate_limit}')
+      except Exception as e:
+        log_error(f'Unable to get github rate limit - {e}')
 
     # Bootstrap repo parameter for bootstrapping
-    if github_bootstrap_repo := params.get('github_bootstrap_repo'):
-      self.bootstrap_repo = self.org.get_repo(f'{github_bootstrap_repo}')
-      log_debug(
-        f'Initialised Github project with bootstrap repo: {self.bootstrap_repo.name}'
-      )
+    if self.bootstrap_repo_name:
+      self.bootstrap_repo = self.org.get_repo(f'{self.bootstrap_repo_name}')
+      log_debug(f'Current bootstrap repo: {self.bootstrap_repo.name}')
     else:
       self.bootstrap_repo = None
 
@@ -61,17 +83,25 @@ class GithubSession:
     log_debug('Authenticating to Github')
     # if the authentication is with a private key, get a fresh token
     if self.private_key:
+      log_debug('Getting a REST access token using the private key')
       try:
         self.rest_token = self.get_access_token()
       except GithubException as g:
         log_critical(f'Unable to authenticate to the github API - {g}')
         sys.exit(1)
     # Then initiate a session with the token
+    log_debug('Getting a token using the REST token')
     self.token = Auth.Token(self.rest_token)
     try:
+      # Then initiate a session with the token
+      log_debug('Starting a session with the token')
       self.session = Github(auth=self.token, pool_size=50)
     except GithubException as e:
       log_critical(f'Unable to create a session using the github API - {e}')
+      sys.exit(1)
+
+    if not self.session:
+      log_critical('Unable to create a session using the github API')
       sys.exit(1)
 
     # Refresh the org object
@@ -101,6 +131,8 @@ class GithubSession:
   def test_connection(self):
     # Test auth and connection to github
     try:
+      if not self.session:
+        return False
       rate_limit = self.session.get_rate_limit()
       self.core_rate_limit = rate_limit.resources.core
       log_info(f'Github API: {rate_limit}')
@@ -276,6 +308,10 @@ class GithubSession:
     return summary
 
   def create_update_pr(self, request):
+    if not self.bootstrap_repo:
+      log_error('Bootstrap repo has not been defined')
+      return request
+
     if request.get('request_type') == 'Archive':
       log_info(f'Request type Archive for repository {request.get("github_repo")}')
       branch_name = f'REQ_ARCHIVE_{request["id"]}_{request.get("github_repo")}'
@@ -400,20 +436,24 @@ class GithubSession:
     return request
 
   def delete_old_workflows(self):
+    if not self.bootstrap_repo:
+      return None
     try:
-      if bootstrap_workflow := [
-        workflow
-        for workflow in self.bootstrap_repo.get_workflows()
-        if workflow.name == 'Bootstrap - poll for repo requests'
-      ]:
-        workflow_runs = bootstrap_workflow[0].get_runs()
-        run_qty = workflow_runs.totalCount
-        if run_qty > 12:
-          log_debug(
-            f'Workflow {bootstrap_workflow[0].name} has {run_qty} runs - cropping to 12'
-          )
-          for run in workflow_runs[12:]:
-            run.delete()
+      if workflows := self.bootstrap_repo.get_workflows():
+        if bootstrap_workflow := [
+          workflow
+          for workflow in workflows
+          if workflow.name == 'Bootstrap - poll for repo requests'
+        ]:
+          workflow_runs = bootstrap_workflow[0].get_runs()
+          run_qty = workflow_runs.totalCount
+          if run_qty > 12:
+            log_debug(
+              f'Workflow {bootstrap_workflow[0].name} has {run_qty} runs -'
+              ' cropping to 12'
+            )
+            for run in workflow_runs[12:]:
+              run.delete()
 
     except GithubException as e:
       log_warning(
